% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\title{softmax function}
\usage{
softmax(x)
}
\arguments{
\item{x}{x is a list}
}
\value{
a list of x's size
}
\description{
It is a simple softmax function.
}
\details{
This is the description of the first function.

you can use this function to caculate and return the softmax values of x.
}
\examples{
\dontrun{
x= c(1,2,3,4)
sx <- softmax(x)
plot(x,sx)
}
}
