% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockPredictionR.R
\name{stock_predict}
\alias{stock_predict}
\title{Predicts Stock PriceMovement for Given Stock Symbol}
\usage{
stock_predict(symbol)
}
\description{
This package predictswhether the stock price at tommorow's market close would be higher or lowercompared to today's closing place.
}
\examples{
\dontrun{
stock_predict('AAPL')
}

}
