% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanh.R
\name{tanh}
\alias{tanh}
\title{tanh function}
\usage{
tanh(x)
}
\arguments{
\item{x}{x is a list, whose element should be best in (-1,1)}
}
\value{
a list of x's size
}
\description{
It is a simple tanh function.
}
\details{
This is the description of the second function.

you can use this function to caculate and return the tanh values of x.
}
\examples{
\dontrun{
x= c(1,2,3,4)
rx <- tanh(x)
plot(x,rx)
}
}
